+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> GRPC Calls for Golang Changelog
 
+## <a name="1.1.0"></a> 1.1.0 (2021-04-04) 
+
+### Breaking Changes
+* Introduced IGrpcServiceOverrides
+* Changed signature NewGrpcService to InheritGrpcService
+* Changed signature NewCommandableGrpcService to InheritGrpcService
+
 ## <a name="1.0.1-1.0.2"></a> 1.0.1-1.0.2 (2020-11-12) 
 
 ### Bug Fixes